[core] Extract global index visibility checker
